requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.0
textual==0.52.1
typing-extensions==4.9.0
uc-micro-py==1.0.3
urllib3==2.2.1
